Feat: enable signAllTransactions for Ingame & External walllets
