--- Runtime/codebase/DeepLinkWallets/PhantomWebGL.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 58053c7a76fb4074afab2d7a87ce762c
-timeCreated: 1664400194
