+++ Runtime/codebase/DeepLinkWallets/Utils.cs
@@ -192,4 +192,57 @@ namespace Solana.Unity.SDK
     {
         public string signature;
     }
+
+        [Serializable]
+    public class PhantomWalletOptions
+    {
+        [SerializeField]
+        private string apiVersion = "v1";
+        public virtual string ApiVersion
+        {
+            get => apiVersion;
+            set => apiVersion = value;
+        }
+
+        [SerializeField]
+        private string appMetaDataUrl = "https://github.com/magicblock-labs/Solana.Unity-SDK";
+        public virtual string AppMetaDataUrl
+        {
+            get => appMetaDataUrl;
+            set => appMetaDataUrl = value;
+        }
+
+        [SerializeField]
+        private string deeplinkUrlScheme = "unitydl";
+        public virtual string DeeplinkUrlScheme
+        {
+            get => deeplinkUrlScheme;
+            set => deeplinkUrlScheme = value;
+        }
+
+        [SerializeField]
+        private string sessionEncryptionPassword = "use a strong password";
+        public virtual string SessionEncryptionPassword
+        {
+            get => sessionEncryptionPassword;
+            set => sessionEncryptionPassword = value;
+        }
+
+        [SerializeField]
+        private string baseUrl = "https://phantom.app";
+        public virtual string BaseUrl
+        {
+            get => baseUrl;
+            set => baseUrl = value;
+        }
+
+        [SerializeField]
+        private string walletName = "phantom";
+        public virtual string WalletName
+        {
+            get => walletName;
+            set => walletName = value;
+        }
+    }
+
 }
