--- Runtime/codebase/PhantomWallet.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 86886b5b62324440870ea00f772204bf
-timeCreated: 1660226458
